requirements.txt
anthropic>=0.40.0
PyGitHub>=2.1.0
slack_sdk>=3.27.0
requests>=2.31.0
APScheduler>=3.10.0,<4.0.0
Flask>=3.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
